requirements.txt
streamlit
pandas
seaborn
matplotlib.pyplot
plotly.graph_objects
pybaseball